+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Multiarch build.
 
 ## [3.0.1] - 2021-11-22
 ### Added